发布合约时生成地址；将工具函数移入util，避免循环依赖
