+++ pbft/client.go
@@ -1,10 +1,8 @@
 package pbft
 
 import (
-	"crypto/rand"
 	"github.com/cloudflare/cfssl/log"
 	"github.com/ssbcV2/common"
-	"math/big"
 )
 
 func ClientSendMessageAndListen() {
@@ -16,18 +14,4 @@ func ClientSendMessageAndListen() {
 	log.Info(" ---------------------------------------------------------------------------------")
 	log.Info("|  已进入PBFT客户端，请启动全部节点后再发送消息！  |")
 	log.Info(" ---------------------------------------------------------------------------------")
-}
-
-//返回一个十位数的随机数，作为msgid
-func getRandom() int {
-	x := big.NewInt(10000000000)
-	for {
-		result, err := rand.Int(rand.Reader, x)
-		if err != nil {
-			log.Error(err)
-		}
-		if result.Int64() > 1000000000 {
-			return int(result.Int64())
-		}
-	}
 }
